+++ src/cache/inmemory/__tests__/policies.ts
@@ -5197,6 +5197,7 @@ describe("type policies", function () {
     function readFirstBookResult() {
       return cache.readQuery<{ author: any }>({
         query: firstBookQuery,
+        canonizeResults: true,
       })!;
     }