+++ src/cache/inmemory/__tests__/readFromStore.ts
@@ -1390,6 +1390,7 @@ describe('reading from the store', () => {
 
   it("propagates eviction signals to parent queries", () => {
     const cache = new InMemoryCache({
+      canonizeResults: true,
       typePolicies: {
         Deity: {
           keyFields: ["name"],
@@ -1879,7 +1880,9 @@ describe('reading from the store', () => {
   });
 
   it("returns === results for different queries", function () {
-    const cache = new InMemoryCache;
+    const cache = new InMemoryCache({
+      canonizeResults: true,
+    });
 
     const aQuery: TypedDocumentNode<{
       a: string[];