+++ src/core/__tests__/QueryManager/index.ts
@@ -8,7 +8,7 @@ import { setVerbosity } from 'ts-invariant';
 
 import { Observable, Observer } from '../../../utilities/observables/Observable';
 import { ApolloLink, GraphQLRequest, FetchResult } from '../../../link/core';
-import { InMemoryCache, InMemoryCacheConfig } from '../../../cache/inmemory/inMemoryCache';
+import { InMemoryCache, InMemoryCacheConfig } from '../../../cache';
 import {
   ApolloReducerConfig,
   NormalizedCacheObject
@@ -833,7 +833,7 @@ describe('QueryManager', () => {
   });
 
   itAsync('will return referentially equivalent data if nothing changed in a refetch', (resolve, reject) => {
-    const request = {
+    const request: WatchQueryOptions = {
       query: gql`
         {
           a
@@ -849,6 +849,7 @@ describe('QueryManager', () => {
         }
       `,
       notifyOnNetworkStatusChange: false,
+      canonizeResults: true,
     };
 
     const data1 = {