+++ src/react/hooks/__tests__/useQuery.test.tsx
@@ -3457,6 +3457,7 @@ describe('useQuery Hook', () => {
   describe('canonical cache results', () => {
     it('can be disabled via useQuery options', async () => {
       const cache = new InMemoryCache({
+        canonizeResults: true,
         typePolicies: {
           Result: {
             keyFields: false,